Cleaning up code related to RISC-V extensions
