+++ security-monitor/src/core/architecture/riscv/mod.rs
@@ -6,7 +6,7 @@ pub use control_status_registers::{ControlStatusRegister, ControlStatusRegisters
 pub use extensions::compressed_instructions::decode_result_register;
 pub use extensions::floating_point_unit::FloatingPointUnit;
 pub use extensions::supervisor_timer_extension::SupervisorTimerExtension;
-pub use extensions::IsaExtension;
+pub use extensions::HardwareExtension;
 pub use general_purpose_registers::{GeneralPurposeRegister, GeneralPurposeRegisters};
 pub use hart_architectural_state::HartArchitecturalState;
 pub use hart_lifecycle_state::HartLifecycleState;
