+++ security-monitor/src/core/architecture/riscv/sbi/spec.rs
@@ -1,0 +1,16 @@
+// SPDX-FileCopyrightText: 2024 IBM Corporation
+// SPDX-FileContributor: Wojciech Ozga <woz@zurich.ibm.com>, IBM Research - Zurich
+// SPDX-License-Identifier: Apache-2.0
+
+pub const SBI_SUCCESS: i16 = 0;
+pub const SBI_ERR_FAILED: i16 = -1;
+pub const SBI_ERR_NOT_SUPPORTED: i16 = -2;
+pub const SBI_ERR_INVALID_PARAM: i16 = -3;
+pub const SBI_ERR_DENIED: i16 = -4;
+pub const SBI_ERR_INVALID_ADDRESS: i16 = -5;
+pub const SBI_ERR_ALREADY_AVAILABLE: i16 = -6;
+pub const SBI_ERR_ALREADY_STARTED: i16 = -7;
+pub const SBI_ERR_ALREADY_STOPPED: i16 = -8;
+pub const SBI_ERR_NO_SHMEM: i16 = -9;
+pub const SBI_ERR_INVALID_STATE: i16 = -10;
+pub const SBI_ERR_BAD_RANGE: i16 = -11;
